# Reduce injector memory allocations