Уменьшено кол-во аллокаций

```
benchstat old.txt new2.txt                          
goos: darwin
goarch: arm64
pkg: github.com/cyberstudio/jsonj
                          │   old.txt    │              new2.txt              │
                          │    sec/op    │   sec/op     vs base               │
EmptyFragmentsGenerator-8   5.256µ ± 16%   4.759µ ± 8%  -9.46% (p=0.000 n=10)

                          │   old.txt    │               new2.txt               │
                          │     B/op     │     B/op      vs base                │
EmptyFragmentsGenerator-8   3.623Ki ± 0%   3.188Ki ± 0%  -12.01% (p=0.000 n=10)

                          │  old.txt   │              new2.txt              │
                          │ allocs/op  │ allocs/op   vs base                │
EmptyFragmentsGenerator-8   60.00 ± 0%   39.00 ± 0%  -35.00% (p=0.000 n=10)
```